Update Node.js CIPD package.
Update Mac and Linux Node.js CIPD packages to 6.10.3 (LTS).

This involves updating the build/upload script to handler the new
package. Because the new Node.js download is distributed as a ".tar.xz"
package, and Python2 doesn't have a built-in ability to handle this, we
update the build script to Python3.

BUG=None
TEST=local
R=iannucci@chromium.org

Change-Id: Ie696b733b281ebaf70e726caaa313633313dd00b
Reviewed-on: https://chromium-review.googlesource.com/497888
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Ryan Tseng <hinoka@chromium.org>