Use "vpython" for Glyco tests.
PYTHONPATH includes "tools/build", which includes a "setuptools" package
that is incompatible with system "pip". This creates a conflict. For
more information, see:
https://chromium-review.googlesource.com/c/498832/

Run Glyco tests through "vpython", which will isolate them in a hermetic
environment and ensure that all of the versions of associated tools are
appropriately matched.

BUG=None
TEST=local

Change-Id: I5c2712519261b6bda216f43a72c8c278fe5ca573
Reviewed-on: https://chromium-review.googlesource.com/498532
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Benjamin Pastene <bpastene@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>