lkgr_finder: Allow overriding lkgr using a forced build
When backing services like milo are down, a forced lkgr build can now be used to easily specify a new lkgr manually.

Since only --manual and not --forced is passed to the lkgr_finder service, the called script will check the revision to be a valid and newer lkgr.

Bug: chromium:720053
Change-Id: Ia1955f99e42c1c8dc91ee2ef3aa3c750afc9c3a0
Reviewed-on: https://chromium-review.googlesource.com/501827
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>