[Monorail] Display ping settings as plain text to users who lack admin perms.
Currently the editing <select> widget is incorrectly shown to users who
do not have permission to edit the field definition.  They cannot submit
the form and it would not be processed if they did, but still, we should
display it as plain text rather than a widget.

Also, I am removing the term "ping" since that word is overused and does
not occur in the actual automated comment or email notification.

Bug: monorail:1519
Change-Id: Iab78d7a83bb21d3d9da76d4649cf68f113158010
Reviewed-on: https://chromium-review.googlesource.com/505150
Reviewed-by: Aneesh Mulye <aneeshm@google.com>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>