[git] Add self-relative PATH overrides.
Currently, the Git wrapper searches for Git by probing the system PATH.
When the Git wrapper is bundled with Git, we want it to always use that
Git regardless of PATH. We accomplish this by configuring the wrapper to
prefer Git in directories relative to its executable.

The current Git deployment places Git in the "bin/" subdirectory
relative to the CIPD root, so we will hard-code that into the wrapper.
On existing systems, no Git will exist at this path, so this will have
no effect.

BUG=chromium:714293
TEST=unit
R=iannucci@chromium.org, nodir@chromium.org

Change-Id: Ifdabe84349fd89d44b981d39aa22de648710bced
Reviewed-on: https://chromium-review.googlesource.com/510702
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>