[Findit] Use try job id if build number not in url
In the case of swarmbucket tryjobs, the url does not contain the
buildnumber. And since this is only used for display purposes, using the
buildbucket id of the try job seems like the way to go. (As that
information is already available to the WfTryjob entity)

R=stgao,lijeffrey,chanli

Bug:
Change-Id: I3c038cc8507f4f83793da30a362abe8fe3cc3cc1
Reviewed-on: https://chromium-review.googlesource.com/511003
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>