[third_party_packages] Disable GECOS in Git.
GECOS is a field in "/etc/passwd" that allows a verbose user name.
Unfortunately, this is (mis-)configured on our bots to include a quote,
and Git identifies that quote as the user name, which promptly fails
a commit author name validity check.

There is little advantage to allowing this obscure mechanism to remain,
both because it is obviously problematic on our fleet and also because
any serious Git committer should have set their own name independently
via "git config".

TBR=iannucci@chromium.org
BUG=chromium:725724, chromium:714852
TEST=prod
  - Ran this in a prod environment that was encountering this issue.
    Issue was no longer present in new package.

Change-Id: I75c76e0bcd66b7960149ae3d9fa0cf54d434d8c7
Reviewed-on: https://chromium-review.googlesource.com/513500
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>