[git] Fix Git wrapper resoltion from symlink.
The Git wrapper will try and resolve CIPD-packaged system Git by using a
directory relative to its executable. However, the current method used
to identify the executable path will resolve symlinks, which will, in
turn, affect the relative directory location.

Change the Git wrapper to safely prefer argv[0] as the path to "self"
instead of the resolved executable path. This will point to the
file/directory that the symlink is in, rather than the executable that
the symlink resolves to, regardless of what "os.Executable" says.

The argv[0] resolution is still checked against the "os.Executable"
resolution via SameFile; argv[0] is only preferred if it is actually the
same file as the resolved executable.

BUG=chromium:714852
TEST=local
R=iannucci@chromium.org

Change-Id: Id53311661bbe6b53385fe50fbee93850d8de6524
Reviewed-on: https://chromium-review.googlesource.com/513526
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>