[Monorail] Fix display of issues with dangling refs.
The problem was that I recently added code to count the number of
IssueRefView objects that are known to be open issues, but that
does not work with dangling issue refs.  They should not be counted.
(A dangling issue reference is a reference to an issue in a project
on code.google.com that was never migrated to monorail.)

Bug: monorail:2712
Change-Id: Ia48455c3e1a62dddfc97a7d6374f508f9f047d54
Reviewed-on: https://chromium-review.googlesource.com/516564
Reviewed-by: Aneesh Mulye <aneeshm@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>