[third_party_packages] Install Git via "copy".
Install the "git" package via CIPD "copy" mode (instead of default,
which is symlink on UNIX platforms).

When Git initializes a new repository, it uses files from its "templates"
directory to seed it by copying those files verbatim. If those files
happen to be symlinks, it will just copy that symlink, at worst dropping it
into a location where it points to junk and at best pointing back to the
ioriginal template, both of which are invalid since they are supposed to
be user-editable. Other uses of templates are similarly broken.

TBR=nodir@chromium.org
BUG=chromium:714852
TEST=expectations,local
  - Reproduced problem locally, confirmed that copy-mode does not
    encounter this problem.

Change-Id: Iebf0f5247bb1b0a3414a59fadf7efdb4902cfca8
Reviewed-on: https://chromium-review.googlesource.com/517354
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>