[Findit] Add missing iteritems() to AssignWarmCacheHost.
Also extending unittest to further check that the dimensions are
correct.

This happenned because by coincidence the only key in the dictionary had
two letters ('id') and therefore it passed the unit_test because
`for k, v in dict` would still work, giving k a value of `i` and v a
value of `d` instead of a value of `id` for k and a value of `slave1`
for v.

R=chanli,stgao,lijeffrey

Bug:
Change-Id: I16d984a99816050b550de3ea8f3850fe37baf4a8
Reviewed-on: https://chromium-review.googlesource.com/520483
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>