luci-migration: do not set Builder bug label
The Builder Monorail bug label was in the initial design where the
luci-migration app searches for bugs in a cron job.
The current implementation files a bug as a part of builder registration,
so there is no need to search for bugs.

Also many values of Builder label increase load on Monorail.

Do not set Builder label.

Bug:
Change-Id: I9417ce3f5f524ddbe6588d4c3205be5bf254f664
Reviewed-on: https://chromium-review.googlesource.com/523438
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>