buildbucket: make canary concept a first-class citizen
Initially the concept of canary was made specific to swarmbucket, but that
abstraction leaked the generic metrics.py. The concept is generic enough that
we can make it a first-class citizen without sacrificing current buildbucket
generality.

Add canary and canary_preference fields to Build entity, see model for more
details. Update the rest of the code to use them.

Bug: 721571
Change-Id: Id271afc17c9edbff794e7aea91530560e5aee358
Reviewed-on: https://chromium-review.googlesource.com/523455
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>