ts_mon: allow for geometric bucketers that don't start with (0, 1)
With the current implementation, it is not possible to have a geometric
bucketer for which the first (non underflow) bucket is not (0, 1).

In some cases, that is quite inconvenient (for example, measuring a time
that is denoted in seconds, but of an event that will frequently take
less than 1 second).

BUG=chromium:729176
TEST=In progress

Change-Id: Idf8359d84627e27daa3a52d000e990dd0f9c10d7
Reviewed-on: https://chromium-review.googlesource.com/523447
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>