[som] Add a stripped-down WCT test runner for running in CI.
This has a nice property of not needing selenium or a Java runtime.

When you run it, it will start up a browser instance via webdriver API,
then tell it to load the test index file, and listen for test results.

It prints client-reported test results to stdout/stderr and exits when
the client has run all of its tests.

Bug:413053
Change-Id: I747437d97da34d629736caef44e07003d22150b9
Reviewed-on: https://chromium-review.googlesource.com/529988
Commit-Queue: Sean McCullough <seanmccullough@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>