[Monorail] Remove special autocomplete logic to only show Restrict labels after typing "r".
That code was meant to hide the complexity of restriction labels from
casual users in most code.google.com projects that never used them
anyway.  For monorail, we focus more on power users, so no reason to
special case this in an inconsistent way.

R=seanmccullough@chromium.org

Bug: monorail:2573
Change-Id: I4ef0d30a794d2fa7c7bdf70817d7f4c4069140bd
Reviewed-on: https://chromium-review.googlesource.com/538114
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>