[eventupload] Fix test race.
Fix a test race in "eventupload". The test sends a tick to a coroutine
and expects the coroutine's output to be in place after the tick is
consumed. However, it doesn't synchronize with an edge that is
guaranteed to occur after "upload" actually finishes. Install a testing
channel to make sure that this edge is hit.

Ensure that each BatchUploader is closed at the end of the test to test
this code path. Run all tests in parallel (since they don't share
globals) both to increase efficiency and expose them to more concurrency
strain.

BUG=chromium:734794
TEST=local
  - Ran with "-parallel 4 -cpu 1,2,4,8 -race", reproduced race without
    change, race no longer manifests with change.

R=katthomas@chromium.org

Change-Id: I2eff98828bb0b2b4419718636c816ef6000e22ce
Reviewed-on: https://chromium-review.googlesource.com/540376
Reviewed-by: Katie Thomas <katthomas@google.com>
Commit-Queue: Daniel Jacques <dnj@chromium.org>