[Findit] Handle tryjobs with no .revision attribute.
It turns out that the .revision attribute of the TryJob tuple is not
what we need to use in most cases.

This change makes it so that when selecting a swarmbucket host we use
the 'bad_revision' property to select the bot.

Also, adding the appropriate test, and changing existing ones to use the
correct assumption.

R=chanli,lijeffrey,wylieb

Bug:
Change-Id: I48bf81bc5d856b962d66afad1453660d379e3995
Reviewed-on: https://chromium-review.googlesource.com/542961
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>