Replace method caching logic with a decorator.
There are many methods in the codebase that repetitively use the exact
same logic to cache the results of their calls. This CL replaces that
logic with the simple ``cached_property`` decorator. This will (1) remove 
the need to re-implement caching logic every time we want to use it, (2) 
improve readability, and (3) remove the need to re-test the caching 
separately on every method that uses it.

Note: I have only implemented this in a single file for now
(clusterfuzz_data.py). After LGTM and submission, I will make another CL 
to expand this one and use ``cached_property`` to replace caching code
everywhere in Predator.

R=katesonia@google.com

Bug:
Change-Id: I91496ac88498bab5c36fd534752f9d7c5943d084
Reviewed-on: https://chromium-review.googlesource.com/538869
Commit-Queue: Cathal Weakliam <cweakliam@google.com>
Reviewed-by: Sharu Jiang <katesonia@chromium.org>