Fixed a non-edit user exception when typing comments.
The prediction code was trying to sugest components,
even to users without permissions to edit or specify
components. It did not null check all of the cases,
so an exception was thrown due to a non-existent element
(the elements that contains the components to edit
or specify).

BUG=monorail:1862

Change-Id: I4fc98274b69e0f8e597e75408e90c639b99f4f2a
Reviewed-on: https://chromium-review.googlesource.com/544382
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>