Fixed a non-edit user issue details page error.
Restricted the inclusion of some edit-specific JavaScript code
from the update form of the issue details page.
It was triggering an exception due to the presumbit running and
trying to change a missing element -
Uncaught TypeError: Cannot read property 'style' of null
    at XMLHttpRequest.onPresubmitResponse (tracker-editing.js:1683)

Also restricted the filter preview section, since it is hidden
for non-edit users.
Both of the restricted sections are now only shown to users with
EditIssue permissions.

BUG=none

Change-Id: Id3a064b0de9c6f079adba5c10e2c7faa1180490b
Reviewed-on: https://chromium-review.googlesource.com/544407
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>