Use --credentials-json in wpt_export recipe.
After some recent changes in the wpt-export script, we are now reading
Gerrit credentials from a file and not just GitHub credentials, so the
argument is being renamed to credentials-json.

I just noticed this in wpt-export and saw that it was a quick change.

Bug: 700092
Change-Id: Iea13d18065a5ba15774127017c0608c655a2f29f
Reviewed-on: https://chromium-review.googlesource.com/548837
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>