In the wpt_import recipe, also pass path to GitHub credentials.
After a recent change, the WPT importer now makes requests to GitHub
just like the exporter does. I copied the GitHub credentials from the
slave running the wpt export script to the slave running the wpt
import script.

The purpose of this CL is to make the importer use the same
credentials as the exporter when making requests to GitHub.

Bug: 727923
Change-Id: Iecbdf38da7b13515de3e76233a6db96695868549
Reviewed-on: https://chromium-review.googlesource.com/555613
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>