[Findit] Compare recent swarmbucket tryjobs vs buildbot.
This change retrieves recent tryjobs done on swarmbucket and matching
runs of the tryjob run on buildbot.

A later CL will provide a user interface to re-run the jobs that do
not have a match on buildbot for the purpose of comparing their
performance.

Bug:
Change-Id: Ib1f802abb67e2eaa03385d8fdad72ead91886d5e
Reviewed-on: https://chromium-review.googlesource.com/554061
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>