Avoided an exception in automatic completion.
When the user typed in a field that supports automatic completion
and then selected a suggestion using the mouse, an exception was
thrown because the code tries to get the caret position of the
active input field after the active input field variable was unset.
Fixed by null checking the active input field variable before
interacting with it.

Bug: none
Change-Id: I99f208ff9c693a151e64bda9da1e6ae635aa676a
Reviewed-on: https://chromium-review.googlesource.com/565659
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>