[Findit] Flake Analyzer - Refactoring FlakeCulprit to separate from MasterFlakeAnalysis
In preparation to send notifications to culprits that introduced flakiness, FlakeCulprit must be it's own entity and keep track of each analysis that suggests it's the culprit. This way only 1 notification will be sent in cases where a CL is responsible for multiple flaky tests.

Bug: 741930
Change-Id: I95bb91162a68fc7103c52f537a18ca64039e49dd
Reviewed-on: https://chromium-review.googlesource.com/571358
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>