[Findit] Add ui for swarmbucket_performance handler.
This change displays the information retrieved by the handler using a
simple table and a couple possible buttons per row.

If the swarmbucket tryjob has a matching buildbot job, the button allows
the user to compare the runs step by step, through a handler that is not
yet implemented. If it does not, a different button allows the user to
trigger such a job, also to a handler that needs to be implemented
still.

R=chanli,lijeffrey,wylieb
BUG=735186

Change-Id: If2dbc21481e43657db646ad8fe3da7ad1d7fba8b
Reviewed-on: https://chromium-review.googlesource.com/560782
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>