Make dataflow/events/common a top level package
When executing a remote run of a dataflow workflow,
the job itself is serialized and sent to the remote worker
(a GCE instance). Nothing else goes with, so if you want
access to local packages, you have to provide the remote
runner with a way to access them. You can either send
the path to a compressed package or send the path to a
setup.py.

I had to move infra/dataflow/events/ to packages/dataflow/ in
order to satisfy infra's requirement to not use relative
imports.

Also add README to packages/dataflow/ and make any modifications
needed to make sure those directions actually work

Bug:615812
Change-Id: I1a185417937e782b105d40828416d940bfc57ac1
Reviewed-on: https://chromium-review.googlesource.com/575653
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Katie Thomas <katthomas@google.com>