[Monorail] Simplify project update methods.
These methods do not need to load the existing project just to see if it
exists, instead checking one value is enough.  And, doing a project load
caused a false positive in our stale cache detection because it is possible,
but not significant, that the project recent_activity_timestamp may have
been updated since the project was loaded at the beginning of processing
this HTTP request.

Also, there is no need to update the fields of the project object in RAM.
The project object in these methods was not shared with any caller or
cached for any later use.

R=seanmccullough@chromium.org

Bug: monorail:2514
Change-Id: I376c74c24bea486f7f11ae50ed01ddf0030cef52
Reviewed-on: https://chromium-review.googlesource.com/578289
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>