[Findit] Fix race condition in buildbucket_client_test.
Since these three tests mock the response to the same url, and the
fetch_url mock is an attribute of the suite (and hence shared between
the testcases) there is a race condition.

By merging these three testcases we avoid this issue, by serializing
access to the mock.

The other test cases in this suite do not have this problem as they
access different URLs.'

R=lijeffrey,chanli,wylieb

Bug:
Change-Id: I8760398280040850fdb9bf795d6f6a4f70cd4785
Reviewed-on: https://chromium-review.googlesource.com/578734
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>