[Monorail] Sort issue.blocking_iids before comparison for debugging.
I am still seeing this comparison fail just due to th order of the
blocking_iids.  The blocking_iids from the database are now in order,
but the existing ones in RAM caches and memcache are sometimes
in a different order.  I looked through every place where blocking_iids
is set, but is is not clear what would change the order, so sorting them
here is much easier.

R=seanmccullough@chromium.org

Bug: monorail:2514
Change-Id: I26e6ddc60d66b8eb7c3008a3d1dff676d18e6861
Reviewed-on: https://chromium-review.googlesource.com/580355
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>