[Predator] Update lines representation in ProfilerStackFrame.
The format of stack frames sent from UMA Sampling Profiler has changed,
and this CL updates the ``ProfilerStackFrame`` class to match. In
particular the profiler now sends a pair of ``lines`` lists (one
for the old version of the code and one for the new), instead of a
single ``callee_lines`` list.

This CL also renames ``CalleeLine`` to ``FunctionLine`` to better
describe its usage.

R=katesonia@google.com

Bug:
Change-Id: I0379d6acd3ed70950c9d53f82d1ddedfc1143cec
Reviewed-on: https://chromium-review.googlesource.com/572642
Commit-Queue: Cathal Weakliam <cweakliam@google.com>
Reviewed-by: Sharu Jiang <katesonia@chromium.org>