[Monorail] Try not using memcache in two-level caches.
I am testing the theory that memcache does not actually lock out writes
or that it is serving stale values after a delete_multi(seconds=5).

If there is still a problem with stale cache entries, it will be simpler
to debug just the RAM cache entries without any dependence on memcache.

If this CL resolves the problem, I will delete the memcache code.  It was
initially intended to prevent high load on the DB when a large number of
new GAE instances are created and all have RAM cache misses.  I think
that our DB servers have enough capacity to handle even that kind of
load spike and mysql maintains its own cache that should be effective
in that situation:
https://dev.mysql.com/doc/refman/5.5/en/query-cache.html

R=agable@chromium.org

Bug: monorail:2514
Change-Id: I3bdd50f0ad9fbbf06da330178974fa534e33bffa
Reviewed-on: https://chromium-review.googlesource.com/582331
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>