[infra_python] Use bundled system Python.
Update the "infra_python" build instructions to use the system-bundled
Python interpreter. This is deployed by Puppet on all build systems.

Currently, "infra_python" CIPD package bakes in the current "infra" ENV.
This includes the prefix of the Python interpreter that was used to
create the environment. This can be problematic when the Python
interpreter that builds the bundle is not present on the systems on
which the bundle is deployed.

Currently, this is mitigated by appending the system's distribution to
the end of the CIPD package name with the expectation that a package
built on a specific distribution will, in turn, have its Python paths
align with the Python on that distribution. This falls apart when the
package is built with non-distribution (e.g., bundled) Python.

To mitigate this, we have deployed a "system Python" CIPD bundle to all
hosts at a fixed path. The "infra_python" bundle will now create a
separate infra VirtualEnv, built against the fixed-path system CIPD bundle,
and include *that* instead of the currently-running Infra VirtualEnv.

Since the system Python bundle is deployed on both builders and
"infra_python" hosts, which will share the same system CIPD bundle
pathing, this is expected to align.

As a plus, because we're using non-native bundles, we can drop the operating
system distribution (and its detection logic) from the "infra_python"
package name!

BUG=chromium:746995
TEST=local

Change-Id: I053e94254c77e84c749bc53164dd7cb7519f39f1
Reviewed-on: https://chromium-review.googlesource.com/582627
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>