[Predator] Make PredatorForUMASamplingProfiler examine renamed paths.
Adds a new ``include_renamed_paths`` flag to the
``TouchCrashedFileMetaFeature``, which makes it also match against the
old version of a file path when a file has been renamed. This flag is
enabled for ``PredatorForUMASamplingProfiler``, since renamed paths are
relevant to some of the Sampling Profiler-detected regressions.

Bug:
Change-Id: Ic84f9ed05556378788c874a31d7cfaf8d565e17e
Reviewed-on: https://chromium-review.googlesource.com/585784
Reviewed-by: Sharu Jiang <katesonia@chromium.org>
Commit-Queue: Cathal Weakliam <cweakliam@google.com>