[Monorail] Avoid DB replication lag when re-applying filter rules.
This CL removes an optimization for re-applying filter rules where the
issues were loaded from DB replicas to avoid putting load on the DB
master.  However, due to replication lag, this could cause stale
data to be read and then written back, causing an issue to overwrite
another change made during the same time or slightly before.

Now, we read from the master.  We use smaller chunks with the same
task execution rate so that it should not place too much load on
the master, and so that there is a smaller critical section.

R=agable@chromium.org

Bug: monorail:2928
Change-Id: I552b9aa510d152760ec8e147031581f0d52a314a
Reviewed-on: https://chromium-review.googlesource.com/592462
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>