[Findit] Allow for tryjobs with no buildbucket response.
In swarmbucket_performance, the way we look for job that are run using
swarmbucket is by querying for all tryjobs in a date range and then
examining their .last_buildbucket_response field. However, this might
not be set for all jobs.

This change takes this into account to avoid an App Engine error.

R=chanli,lijeffrey,wylieb,stgao
BUG=750364

Change-Id: I4883c326848299266330a616fe1a475df679e31e
Reviewed-on: https://chromium-review.googlesource.com/594359
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>