[Monorail] Update call to CreateIssueComment for inbound alerts.
I had to edit this code to match a change that I made recently to
CreateIssueComment().  In updating it, I realized that it needed a
unit test and that there seemed to be an indentation error that
affected the logic of searching for the most recently modified
issue with the given incident label.

R=zhangtiff@chromium.org

Bug: monorail:2514
Change-Id: I171434e63496f1ce7ffc50e6382189c68dcfa07a
Reviewed-on: https://chromium-review.googlesource.com/594928
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>