[Predator] Make TouchCrashedDirectory ignore test files.
The ``TouchCrashedDirectoryFeature`` very frequently triggers because of
changes in test files. These changes are irrelevant and just add
noise, so this CL adds a flag to allow for ignoring them. I only ignore
them in ``PredatorForUMASamplingProfiler`` for now, since this is the
only version I have tested it on with real data.

R=katesonia@google.com

Bug:
Change-Id: Ie95e5309a2379d5c28fdc2f81089025287170482
Reviewed-on: https://chromium-review.googlesource.com/592421
Commit-Queue: Cathal Weakliam <cweakliam@google.com>
Reviewed-by: Sharu Jiang <katesonia@chromium.org>