[Predator] Handle ``FileChangeInfo`` paths better.
I noticed Predator crash a couple of times while running a delta test,
due to it trying to use the ``new_path`` field of a ``FileChangeInfo``
when the ``change_type`` was 'delete'. In this case the ``new_path``
field is ``None``. This usually didn't matter, but under the right
circumstances it would cause an error, so now I handle this case
explicitly to prevent any problems.

R=katesonia@google.com

Bug:
Change-Id: I5f0ad919a966a1dfd63dcc1ce9ed873d4984f066
Reviewed-on: https://chromium-review.googlesource.com/599151
Reviewed-by: Sharu Jiang <katesonia@chromium.org>
Commit-Queue: Cathal Weakliam <cweakliam@google.com>