[bq] BigQuery TableDef proto uses dataset enum.
Currently, the "DatasetId" proto field is implicitly constrainted to a
handful of supported dataset values. Formally encode this as an
enumeration in the protobuf, and update associated tooling and libraries
to support this.

Update "bqexport" to output to a base directory rather than a full
output path. It will then use the dataset ID as a subdirectory so that
it maintains the layout currently in place by "bqschemaupdater".

BUG=chromium:742478
TEST=None

Change-Id: I5f64b55eaf94beffe3d7bdb64e84390d052f4843
Reviewed-on: https://chromium-review.googlesource.com/619281
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Katie Thomas <katthomas@google.com>