Add LegacyCompletedBuild/Steps schema.
This is a modified version of the schemas used by BuildBot builds.
Those schemas are full of unused or useless fields. This schema trims
those fields and adds some additional fields.

The schemas used here are deemed "legacy". They are propagated
into BigQuery event monitoring because we want to have general
compatibility with the current BuildBot build events, but the intent
is to revise them significantly and export a new table that takes better
advantage of BigQuery's schema layout and is more suited to our expanding
infrastructure, notably LUCI.

BUG=chromium:742478
TEST=None

Change-Id: I779da54f9aa59ab1954e2866906c48d0b3d837a9
Reviewed-on: https://chromium-review.googlesource.com/614565
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Katie Thomas <katthomas@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>