[Predator] Improve MinDistance feature for sampling profiler.
Expands the ``crashed_line_numbers`` field of ``ProfilerStackFrame`` to
also include function call lines. The purpose of this change is to give
more information for the ``MinDistance`` feature. Based on testing with
real data, this improves the accuracy of the feature in many cases.
(However the impact on the model's overall accuracy is negligible.)

R=katesonia@google.com

Bug:
Change-Id: I09eed29a06bfe96449130ebcd337be740c531823
Reviewed-on: https://chromium-review.googlesource.com/619299
Reviewed-by: Sharu Jiang <katesonia@chromium.org>
Commit-Queue: Cathal Weakliam <cweakliam@google.com>