[bbroll] Allow rolling of non-infra.git packages.
In particular, this works for the 'git' third-party package now, but will
also support the 'python' third-party package (when we start using it), as
well as anything else. It assumes:
  * The package is tagged with a `version:<something>` tag
  * The package maintains a `latest` ref
  * It uses the current machine's expansion of ${platform} when searching
    CIPD for the latest.

This also renames the `git` pin to `git_wrapper` to reflect reality, and
refactors a bunch of methods to be inside of PinConfig, because things
were getting a bit dicey with the allowance of pins working with and
without infra.git.

R=dnj@chromium.org, nodir@chromium.org

Bug:
Change-Id: I8553c1a68f6d05cca25aa6747026fab033419ec1
Reviewed-on: https://chromium-review.googlesource.com/620190
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>