android_docker: Lock on individual devices and break into threads.
This is to prevent a device that is constantly disconnecting and
reconnecting from hogging the flock and starving out other invocations.

This also breaks the initial usb bus querying into its own flock. This
is still a bottleneck for all invocations but I'm hoping it's quick
and painless enough to remain. It's also possible that it's unneeded,
but better safe than sorry.

Bug: 670879
Change-Id: I9e7044351c18af2898e4d111c0913f1552e9c408
Reviewed-on: https://chromium-review.googlesource.com/634514
Commit-Queue: Benjamin Pastene <bpastene@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>