[Findit] Flake Analyzer - DataPoints have None for iterations.
The function EstimateSwarmingIterationTimeout examines the analysis for
data points and uses the datapoints + swarming tasks to estimate a
a timeout for one iteration.

For whatever reason some datapoints have None in the place of
iterations.

This change makes EstimateSwarmingIterationTimeout defensive, although
the assumption that FlakeSwarmingTask is well-formed is still there.

Bug:760812
Change-Id: If1d4ff4dc9146c78ac6ed7cc7f77e3edce4ab911
Reviewed-on: https://chromium-review.googlesource.com/644726
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>