[Predator] Modify filtering logic for sampling profiler.
In cases where a performance change is a shift in execution at the root
(due for example to a function being added, deleted, or renamed), then
don't filter out the frame above the root. This frame usually has very
useful information for CL attribution, since the culprit CL will have
changed the exact line of the function call in that frame.

Based on discussions with @wittman, the long-term plan is to simplify
the filtering logic to just include any frames where the 'responsible'
field is True. That way, changes to filtering logic can easily be made
in the regression detection pipeline by just changing how the
'responsible' field is set. However, this is the short term solution
until the necessary changes can be made in the pipeline, and the old
data can be re-generated.

R=katesonia@google.com

Bug:
Change-Id: I7b1255b82652b67973d5806673ecfe6be4ae4a5b
Reviewed-on: https://chromium-review.googlesource.com/644868
Reviewed-by: Sharu Jiang <katesonia@chromium.org>
Commit-Queue: Sharu Jiang <katesonia@chromium.org>