[Findit] Flake Analyzer - Do not use exit codes when determining swarming task success
When swarming returns exit_code, it is the underlying process' exit code which is arbitrarily
chosen and should not be relied on when determining whether a test is stable, flaky, or some
error occurred. In order to determine a test's outcome more reliably, the output_json must be
parsed.

1. The task's status must be 'COMPLETED'
2. 'outputs_ref' Must be available in the swarming task result
3. The task's failure log must be available
4. 'per_iteration_data' Must be available and non-empty in the task's failure log.

When all criteria are met can a task's outcome be determined. If any criteria is not met, assume
there was an error in the task.

Bug: 700980
Change-Id: I947cef9444a8a7605bdc0f9854ab02835c13c98f
Reviewed-on: https://chromium-review.googlesource.com/644428
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>