[Findit] Flake Analyzer - Stabilize dtpr pipeline.
This addresses some problems with the determine true pass rate pipeline.
(1) pipeline.After doesn't work like I thought it did so changing the 
flow a bit to accommodate that.
(2) Added a lower threshold of iterations that must be reached in order
to be considered for convergence.
Bug:762190
Change-Id: Icb021686106749c07db1089939ede2cf1cb16ac8
Reviewed-on: https://chromium-review.googlesource.com/651078
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@chromium.org>