[Monorail] Outline a big refactoring to move business logic into a shared code path.
In this step, I am creating the WorkEnv class and some methods as
examples, along with tests and updates to some callers.  The progress
here is that the caller code gets simplified and there is a place to
move permission checking and rate limiting to in upcoming CLs.

R=seanmccullough@chromium.org

Bug: monorail:2985
Change-Id: I8723384a02edcf2824d7b6d480ec33654ab10fff
Reviewed-on: https://chromium-review.googlesource.com/621386
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>