[Findit] Fix bug when analysis current failure.
Because of my last change https://chromium-review.googlesource.com/c/infra/infra/+/620008, analysis on the current failed build is before build_failure_analysis_pipelines.NeedANewAnalysis. But since NeedANewAnalysis is where we create new WfAnalysis entities
for new analysis, this change introduced a bug where we use the analysis entity when it's not created yet.

This CL fixes the bug by analyzing on current failed build after NeedANewAnalysis (which means entity is created or reset), and still
keep the functionality to bail out if the failure is not really failed or some required information is missing.

Bug: 762704
Change-Id: Ic1ebbcfd6ba8faacc4884633b012600b67911c20
Reviewed-on: https://chromium-review.googlesource.com/654065
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>