[Monorail] Simplify cache registraion and fix invalidation of sharded ram cache.
The current cache structure looks like:
service_object -> TableSpecificTwoLevelCache -> RamCache <-> CacheManager
There is an assumption that the cache keys will be integers, 
such as user ID, project ID, or issue ID.  In this case, 
project ID is assumed.

The LabelRowTwoLevelCache has logic to use 
(project_id, shard_id) instead of just project_id.  However, that
modification is not present in the all-purpose RamCache class, so 
it is not considered for distributed invalidations.

In this CL, I add a class ShardedRamCache that that maps from the
integer-only Invalidate ID numbers to all the (project_id, shard_id)
keys used for that project in the ram cache.

R=agable@chromium.org

Bug: monorail:3026
Change-Id: I1f22e6f4273c90474221b2dd2540ca066cc30a8d
Reviewed-on: https://chromium-review.googlesource.com/656514
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>