[ts_mon]: Move super() calls at end of __init__
This is necessary because the top of the inheritance chain, Metric.__init__,
calls interface.register. register() requires the metric to be equal to the
existing metric of the same name (if it exists), so by the time Metric.__init__
is called, we have to have already finished up the initialization we were going
to do in the subclass (e.g. setting .bucketer in _DistributionBase.__init__).

See https://chromium-review.googlesource.com/c/556465

BUG=chromium:730262

Change-Id: I4bb6f0002a089209e8c94618de529a38705ba009
Reviewed-on: https://chromium-review.googlesource.com/663888
Reviewed-by: Dave Sansome <dsansome@chromium.org>
Commit-Queue: Paul Hobbs <phobbs@google.com>