SOM: Group alerts using annotations
This CL adds a post processing step to the current analysis pipeline.
This processes all known alerts, and groups them according to "reason".
This was already done in the existing logic; we're mostly moving this
to the SOM code so that we can use the grouping logic present in
annotations, which gives users more flexibility if the automatic groupings
are inaccurate or need to be corrected.

Change-Id: I1f38c9cc80ba6f906ecbd8d912bc6e817c6d382f
Reviewed-on: https://chromium-review.googlesource.com/633871
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>