[Predator] Change the way of matching component.
Previously, we order the components and return the first matched component.
This way is problematic because we order them by the longest directory path. So
it is possible that:

    {
      "component": "Blink>JavaScript",
      "dirs": [
        "src/v8",
        "src/v8/src/base/blabla..."
      ]
    },

is ranked higher than:

    {
      "component": "Blink>JavaScript>GC",
      "dirs": [
        "src/v8/src/heap"
      ]
    },

If we have file_path like 'src/v8/src/heap/a.cc', it will be matched to "Blink>JavaScript" instead of "Blink>JavaScript>GC".

So, I changed the behavior to let the file_path match all components, and return the component with longest (most '/') matched directory path.

For the example above, both the "Blink>JavaScript" and "Blink>JavaScript>GC" are matched component, but the matched directory for "Blink>JavaScript>GC" - 'src/v8/src/heap' has more '/', so we should return "Blink>JavaScript>GC".

Bug: 766392
Change-Id: I3467c3cca7dd3738e6733b03039451ecd015ccf8
Reviewed-on: https://chromium-review.googlesource.com/672004
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>