[Findit] Flake Analyzer - Rework timeouts for dtpr pipeline.
Currently the dtpr has a target iterations that it wants to get
run, and uses that as a basis of calculation. The strategy is
flipped where it has a certain time period it wants to fill
and calculates the number of tests that it needs to fill it.
Along with this, now that flake swarming tasks are being deleted
at the start of every new task, I added elpased second to the
flake analysis model to  fill in the approximation of time.

Bug:764961
Change-Id: Ib3b033f48c2b3f7ac72d1b5973aa2bd9876d9963
Reviewed-on: https://chromium-review.googlesource.com/666320
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>