[Predator] Add repo_to_dep_path to CrashConfig.
test url: https://katesonia-test-test-dot-findit-for-me.googleplex.com/config

The repo_to_dep_path is for us to map a repository url to its chromium based dep path, so we can correctly match paths from stand-alone component build to chromium paths.

In the directory owner mapping https://storage.googleapis.com/chromium-owners/component_map.json (What we use to find a component.), the paths are all relative to the chromium repo. 

For example,
For those d8 build like this:
https://findit-for-me.googleplex.com/clusterfuzz/result-feedback?key=ahpzfmdvb2dsZS5jb206ZmluZGl0LWZvci1tZXJBCxITQ2x1c3RlcmZ1enpBbmFseXNpcyIoMjA4YWY4ZTkwOGEwYTk3YmZiOThlMjliNmYzZDUxZjA1YjNjMmMzZgw

We build from v8 repo, and we don't know the Chromium DEPS information. We should do a stand-alone repo path -> chromium repo path conversion.
The dep_path is v8, we should map it to src/v8, another dep of v8 is v8/third_party/tools, we should map it to src/third_party/tools. This conversion is only for finding a component, since the component_map.json keeps being updated, so we use the same latest DEPS. (If the dependency in DEPS at the revision is not in the master DEPS, then we won't find a match in component_map.json anyway)

Bug:
Change-Id: Id8ad893f2063c7228ffc0b7fcfe08d35d848fe86
Reviewed-on: https://chromium-review.googlesource.com/677565
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Sharu Jiang <katesonia@chromium.org>